Fix automation scripts and configuration files
