+++ automation-logs.txt
@@ -1325,3 +1325,31 @@ This file is being treated as an ES module because it has a '.js' file extension
 [2025-09-07T02:15:48.062Z] Running: Git Status
 [2025-09-07T02:15:48.101Z] ✅ Git Status completed successfully
 [2025-09-07T02:15:48.101Z] Running: Git Add
+[2025-09-07T02:15:48.157Z] ✅ Git Add completed successfully
+[2025-09-07T02:15:48.158Z] Running: Git Commit
+[2025-09-07T02:15:48.212Z] ✅ Git Commit completed successfully
+[2025-09-07T02:15:48.212Z] Running: Git Push
+[2025-09-07T02:15:49.052Z] Command failed: git push origin main - Command failed: git push origin main
+To https://github.com/Zion-Holdings/zion.app
+ ! [rejected]                main -> main (fetch first)
+error: failed to push some refs to 'https://github.com/Zion-Holdings/zion.app'
+hint: Updates were rejected because the remote contains work that you do not
+hint: have locally. This is usually caused by another repository pushing to
+hint: the same ref. If you want to integrate the remote changes, use
+hint: 'git pull' before pushing again.
+hint: See the 'Note about fast-forwards' in 'git push --help' for details.
+
+[2025-09-07T02:15:49.052Z] ❌ Git Push failed: To https://github.com/Zion-Holdings/zion.app
+ ! [rejected]                main -> main (fetch first)
+error: failed to push some refs to 'https://github.com/Zion-Holdings/zion.app'
+hint: Updates were rejected because the remote contains work that you do not
+hint: have locally. This is usually caused by another repository pushing to
+hint: the same ref. If you want to integrate the remote changes, use
+hint: 'git pull' before pushing again.
+hint: See the 'Note about fast-forwards' in 'git push --help' for details.
+
+[2025-09-07T02:15:49.052Z] 📊 Report saved to /workspace/automation/logs/comprehensive-automation-report.json
+[2025-09-07T02:15:49.052Z] 🎉 Comprehensive automation run completed
+[2025-09-07T02:15:49.052Z] ✅ Successful: 11
+[2025-09-07T02:15:49.052Z] ❌ Failed: 5
+[2025-09-07T02:15:49.053Z] ⚠️ Warnings: 0
