+++ next.config.cjs
@@ -1,53 +1,79 @@
-
-
-  reactStrictMode: true,
-  eslint: {
-    ignoreDuringBuilds: true
-  },
-  typescript: {
-    ignoreBuildErrors: true
-  },
-  experimental: {
-    esmExternals: false
-  },
-  images: {
-    domains: ['ziontechgroup.com', 'localhost', 'images.unsplash.com', 'via.placeholder.com'],
-    unoptimized: true
-  },
-    ignoreDuringBuilds: true,
-  reactStrictMode: false,
-  swcMinify: false,
-/** @type {import('next').NextConfig} */
 const nextConfig = {
-  reactStrictMode: true,
-  swcMinify: true,
   compress: true,
   poweredByHeader: false,
-  eslint: { ignoreDuringBuilds: true },
-  typescript: { ignoreBuildErrors: true },
+  generateEtags: true,
+  
+  // Disable linting during build
+  eslint: {
+    ignoreDuringBuilds: true,
+  },
+  
+  // Disable type checking during build
+  typescript: {
+    ignoreBuildErrors: true,
+  },
+  
+  // Image optimization
   images: {
-    domains: ["localhost", "ziontechgroup.com"],
+    domains: ['localhost'],
     formats: ['image/webp', 'image/avif'],
-    deviceSizes: [640, 750, 828, 1080, 1200, 1920, 2048, 3840],
-    imageSizes: [16, 32, 48, 64, 96, 128, 256, 384]
   },
-  experimental: {
-    // Remove optimizePackageImports to satisfy Next 12 validation
-  },
+  
+  // Security headers
   async headers() {
     return [
       {
         source: '/(.*)',
         headers: [
-          { key: 'X-Content-Type-Options', value: 'nosniff' },
-          { key: 'X-Frame-Options', value: 'DENY' },
-          { key: 'X-XSS-Protection', value: '1; mode=block' },
-          { key: 'Referrer-Policy', value: 'origin-when-cross-origin' }
-        ]
-      }
+          {
+            key: 'X-Frame-Options',
+            value: 'DENY',
+          },
+          {
+            key: 'X-Content-Type-Options',
+            value: 'nosniff',
+          },
+          {
+            key: 'Referrer-Policy',
+            value: 'origin-when-cross-origin',
+          },
+        ],
+      },
     ];
-  }
+  },
+  
+  // Redirects
+  async redirects() {
+    return [
+      {
+        source: '/home',
+        destination: '/',
+        permanent: true,
+      },
+    ];
+  },
+  
+  // Webpack configuration
+  webpack: (config, { buildId, dev, isServer, defaultLoaders, webpack }) => {
+    // Add custom webpack configuration here if needed
+    return config;
+  },
+  
+  // Experimental features
+  experimental: {
+    optimizeCss: true,
+  },
+  
+  // Output configuration
+  output: 'standalone',
+  
+  // Trailing slash
+  trailingSlash: false,
+  
+  // Environment variables
+  env: {
+    CUSTOM_KEY: process.env.CUSTOM_KEY,
+  },
 };
 
 module.exports = nextConfig;
-
