--- next.config.js
@@ -1,79 +1,0 @@
-const nextConfig = {
-  compress: true,
-  poweredByHeader: false,
-  generateEtags: true,
-  
-  // Disable linting during build
-  eslint: {
-    ignoreDuringBuilds: true,
-  },
-  
-  // Disable type checking during build
-  typescript: {
-    ignoreBuildErrors: true,
-  },
-  
-  // Image optimization
-  images: {
-    domains: ['localhost'],
-    formats: ['image/webp', 'image/avif'],
-  },
-  
-  // Security headers
-  async headers() {
-    return [
-      {
-        source: '/(.*)',
-        headers: [
-          {
-            key: 'X-Frame-Options',
-            value: 'DENY',
-          },
-          {
-            key: 'X-Content-Type-Options',
-            value: 'nosniff',
-          },
-          {
-            key: 'Referrer-Policy',
-            value: 'origin-when-cross-origin',
-          },
-        ],
-      },
-    ];
-  },
-  
-  // Redirects
-  async redirects() {
-    return [
-      {
-        source: '/home',
-        destination: '/',
-        permanent: true,
-      },
-    ];
-  },
-  
-  // Webpack configuration
-  webpack: (config, { buildId, dev, isServer, defaultLoaders, webpack }) => {
-    // Add custom webpack configuration here if needed
-    return config;
-  },
-  
-  // Experimental features
-  experimental: {
-    optimizeCss: true,
-  },
-  
-  // Output configuration
-  output: 'standalone',
-  
-  // Trailing slash
-  trailingSlash: false,
-  
-  // Environment variables
-  env: {
-    CUSTOM_KEY: process.env.CUSTOM_KEY,
-  },
-};
-
-module.exports = nextConfig;
